Change all dataclasses to manual classes
