+++ ionerdss/model/components.py
@@ -28,12 +28,13 @@ Dependencies:
 """
 
 import json
-import numpy as np
-from dataclasses import dataclass, field
+from dataclasses import field
 from typing import List, Tuple
-from ..math.coords import Coords
 
-@dataclass
+import numpy as np
+
+from ionerdss.math.coords import Coords
+
 class MoleculeInterface:
     """Represents an interface of a molecule type.
     
@@ -41,53 +42,91 @@ class MoleculeInterface:
         name (str): Name of the interface.
         coord (Tuple[float, float, float]): Coordinates of the interface.
     """
-    name: str
-    coord: Coords
+    def __init__(self, name, coord):
+        self.name = name
+        self.coord = coord
 
-@dataclass
+    def __repr__(self):
+        return f"<MoleculeInterface {self.name} @ {self.coord}>"
+
 class MoleculeType:
-    """Represents a molecule type in the model.
-    
-    Attributes:
-        name (str): Name of the molecule type.
-        interfaces (List[MoleculeInterface]): List of interfaces associated with the molecule.
     """
-    name: str
-    interfaces: List[MoleculeInterface]
-    diffusion_translation: float = 0.0
-    diffusion_rotation: float = 0.0
+    Represents a molecule type in the model.
 
-@dataclass
+    Attributes
+    ----------
+    name : str
+        Name of the molecule type.
+    interfaces : list of MoleculeInterface
+        List of interfaces associated with the molecule.
+    diffusion_translation : float
+        Translational diffusion constant (default 0.0).
+    diffusion_rotation : float
+        Rotational diffusion constant (default 0.0).
+    """
+    def __init__(self, name, interfaces=None, diffusion_translation=0.0, diffusion_rotation=0.0):
+        self.name = name
+        self.interfaces = interfaces if interfaces is not None else []
+        self.diffusion_translation = diffusion_translation
+        self.diffusion_rotation = diffusion_rotation
+
+    def __repr__(self):
+        return f"<MoleculeType {self.name} with {len(self.interfaces)} interfaces>"
+
 class Reaction:
-    """Represents a reaction in the model.
-    
-    Attributes:
-        name (str): Reaction expression.
-        binding_radius (float): Binding radius of the reaction.
-        binding_angles (Tuple[float, float, float, float, float]): Binding angles.
-        norm1 (Tuple[float, float, float]): Normal vector 1.
-        norm2 (Tuple[float, float, float]): Normal vector 2.
     """
-    name: str
-    binding_radius: float
-    binding_angles: Tuple[float, float, float, float, float]
-    norm1: Tuple[float, float, float]
-    norm2: Tuple[float, float, float]
-    ka: float = 0.0  # Forward rate constant
-    kb: float = 0.0  # Reverse rate constant
+    Represents a reaction in the model.
 
-@dataclass
-class Model:
-    """Parent class for all models to generate input files for NERDSS simulations.
-    
-    Attributes:
-        name (str): Name of the model.
-        molecule_types (List[MoleculeType]): List of molecule types in the model.
-        reactions (List[Reaction]): List of reactions in the model.
+    Attributes
+    ----------
+    name : str
+        Reaction expression.
+    binding_radius : float
+        Binding radius of the reaction.
+    binding_angles : tuple of float
+        Binding angles (theta1, theta2, phi1, phi2, chi).
+    norm1 : tuple of float
+        First normal vector.
+    norm2 : tuple of float
+        Second normal vector.
+    ka : float
+        Forward rate constant.
+    kb : float
+        Reverse rate constant.
     """
-    name: str
-    molecule_types: List[MoleculeType] = field(default_factory=list)
-    reactions: List[Reaction] = field(default_factory=list)
+    def __init__(self, name, binding_radius, binding_angles,
+                 norm1, norm2, ka=0.0, kb=0.0):
+        self.name = name
+        self.binding_radius = binding_radius
+        self.binding_angles = binding_angles
+        self.norm1 = norm1
+        self.norm2 = norm2
+        self.ka = ka
+        self.kb = kb
+
+    def __repr__(self):
+        return f"<Reaction {self.name}, ka={self.ka}, kb={self.kb}>"
+
+class Model:
+    """
+    Parent class for all models to generate input files for NERDSS simulations.
+
+    Attributes
+    ----------
+    name : str
+        Name of the model.
+    molecule_types : list of MoleculeType
+        List of molecule types in the model.
+    reactions : list of Reaction
+        List of reactions in the model.
+    """
+    def __init__(self, name, molecule_types=None, reactions=None):
+        self.name = name
+        self.molecule_types = molecule_types if molecule_types is not None else []
+        self.reactions = reactions if reactions is not None else []
+
+    def __repr__(self):
+        return f"<Model {self.name}: {len(self.molecule_types)} molecules, {len(self.reactions)} reactions>"
 
     def save_model(self, file_path: str) -> None:
         """Saves the model to a specified JSON file.
@@ -121,11 +160,10 @@ class Model:
                 for rxn in self.reactions
             ],
         }
-        with open(file_path, "w") as file:
+        with open(file_path, "w", encoding="utf-8") as file:
             json.dump(data, file, indent=4, cls=CustomJSONEncoder)  # Use custom encoder
 
-    @classmethod
-    def load_model(cls, file_path: str) -> "Model":
+    def load_model(self, cls, file_path: str) -> "Model":
         """Loads a model from a specified JSON file.
         
         Args:
@@ -164,12 +202,216 @@ class Model:
 
 class CustomJSONEncoder(json.JSONEncoder):
     """Custom JSON encoder to handle Coords serialization and NumPy types."""
+    def default(self, o):
+        if isinstance(o, Coords):
+            return {"x": o.x, "y": o.y, "z": o.z}  # Convert Coords to dict
+        elif isinstance(o, np.float32):  # Convert numpy float32 to standard float
+            return float(o)
+        elif isinstance(o, np.ndarray):  # Convert numpy array to list
+            return o.tolist()
+        return super().default(o)
+
+class MoleculeTemplate:
+    """
+    Represents a molecule type in NERDSS, including the molecule's center of mass (COM) 
+    and a list of binding interfaces.
+
+    Attributes:
+        name (str): Identifier for the molecule type.
+        interface_template_list (list): A list of BindingInterfaceTemplate objects that 
+            describe the molecule’s binding sites.
+        normal_point (list): Default normal vector direction.
+    """
+
+    def __init__(self, name: str):
+        """
+        Initializes a MoleculeTemplate.
+
+        Args:
+            name (str): Name/identifier of the molecule type.
+        """
+        self.name = name
+        self.interface_template_list = []
+        self.normal_point = [0,0,1]
+        self.diffusion_translation = None
+        self.diffusion_rotation = None
+        self.radius = None
+
+    def __str__(self):
+        interfaces = "\n  ".join(str(it) for it in self.interface_template_list)
+        return f"Molecule Template: {self.name}\n  Interfaces:\n  {interfaces}"
     
-    def default(self, obj):
-        if isinstance(obj, Coords):
-            return {"x": obj.x, "y": obj.y, "z": obj.z}  # Convert Coords to dict
-        elif isinstance(obj, np.float32):  # Convert numpy float32 to standard float
-            return float(obj)
-        elif isinstance(obj, np.ndarray):  # Convert numpy array to list
-            return obj.tolist()
-        return super().default(obj)
+    def __eq__(self, other):
+        if not isinstance(other, MoleculeTemplate):
+            return False
+        return self.name == other.name
+
+class InterfaceTemplate:
+    """
+    Represents a binding interface template between molecules.
+
+    Attributes:
+        name (str): Identifier of the interface template.
+        coord (Coords): Relative coordinates of the interface.
+        my_residues (list): Residues forming this interface.
+        required_free_list (list): Other interface templates that must remain unbound 
+            for this interface to bind.
+        signature (dict): Stores interface geometry information.
+    """
+
+    def __init__(self, name: str):
+        """
+        Initializes a BindingInterfaceTemplate.
+
+        Args:
+            name (str): Identifier for the interface template.
+        """
+        self.name = name
+        self.coord = None
+        self.my_residues = []
+        self.required_free_list = [] # The list of interface templates that need to be free to bind to this interface template
+        self.signature = {}
+        self.energy = None
+
+    def __str__(self):
+        residues = ", ".join(self.my_residues)
+        required_free = ", ".join(self.required_free_list)
+        return (f"Interface Template: {self.name}\n"
+                f"  Coordinates: {self.coord}\n"
+                f"  Residues: {residues}\n"
+                f"  Required Free: {required_free}")
+    
+    def __eq__(self, other):
+        if not isinstance(other, InterfaceTemplate):
+            return False
+        # TODO: check this
+        return self.name == other.name
+
+class CoarseGrainedMolecule:
+    """
+    Represents a coarse-grained molecule in NERDSS, potentially derived from a PDB chain.
+
+    Attributes:
+        name (str): Identifier of the molecule.
+        my_template (MoleculeTemplate): Reference to the associated molecule template.
+        coord (Coords): Center-of-mass coordinates.
+        interface_list (list): List of binding interfaces.
+        normal_point (list): Normal vector direction.
+    """
+
+    def __init__(self, name: str):
+        """
+        Initializes a CoarseGrainedMolecule.
+
+        Args:
+            name (str): Name/identifier of the molecule.
+        """
+        self.name = name
+        self.my_template = None
+        self.coord = None
+        self.interface_list = []
+        self.normal_point = None
+        self.diffusion_translation = None
+        self.diffusion_rotation = None
+        self.radius = None
+
+    def __str__(self):
+        interfaces = "\n  ".join(str(interface) for interface in self.interface_list)
+        return (f"CoarseGrainedMolecule: {self.name}\n"
+                f"  Template: {self.my_template}\n"
+                f"  Coordinates: {self.coord}\n"
+                f"  Interfaces:\n  {interfaces}")
+    
+    def __repr__(self):
+        # Similar to __str__ but more formal for debugging
+        return self.name
+    
+    def __eq__(self, other):
+        if not isinstance(other, CoarseGrainedMolecule):
+            return False
+        return self.name == other.name
+    
+    def __hash__(self):
+        return hash(self.name)
+
+class BindingInterface:
+    """
+    Represents a binding interface between molecules.
+
+    Attributes:
+        name (str): Identifier of the binding interface.
+        coord (Coords): Position of the interface.
+        my_template (BindingInterfaceTemplate): Reference to the associated interface template.
+        my_residues (list): Residues included in the interface.
+        signature (dict): Stores interface geometry information.
+    """
+
+    def __init__(self, name: str):
+        """
+        Initializes a BindingInterface.
+
+        Args:
+            name (str): Identifier for the binding interface.
+        """
+        self.name = name
+        self.coord = None
+        self.my_template = None
+        self.my_residues = []
+        self.signature = {}
+        self.energy = None
+
+    def __str__(self):
+        return (f"BindingInterface: {self.name}\n"
+                f"  Template: {self.my_template}\n"
+                f"  Coordinates: {self.coord}\n"
+                f"  Residue Count: {len(self.my_residues)}\n"
+                f"  Residues: {self.my_residues}")
+    
+    def __eq__(self, other):
+        if not isinstance(other, BindingInterface):
+            return False
+        return self.my_template == other.my_template
+
+class ReactionTemplate:
+    """
+    Defines a reaction template between two MoleculeTemplates.
+
+    Attributes:
+        expression (str): Textual representation of the reaction.
+        reactants (list): List of reactant molecule/interface templates.
+        products (list): List of product molecule/interface templates.
+        binding_angles (tuple): Tuple describing binding angles (theta1, theta2, phi1, phi2, omega).
+        binding_radius (float): Distance between binding interfaces.
+        norm1 (list): Normal vector of the first reactant.
+        norm2 (list): Normal vector of the second reactant.
+    """
+
+    def __init__(self):
+        """
+        Initializes a ReactionTemplate with default values.
+        """
+        self.expression = None
+        self.reactants = None
+        self.products = None
+        self.binding_angles = None
+        self.binding_radius = None
+        self.norm1 = None
+        self.norm2 = None
+        self.kd = None
+        self.ka = None
+        self.kb = None
+        self.energy = None
+
+    def __str__(self):
+        return (f"Reaction Template: {self.expression}\n"
+                f"  Reactants: {self.reactants}\n"
+                f"  Products: {self.products}\n"
+                f"  Binding Angles: {self.binding_angles}\n"
+                f"  Binding Radius: {self.binding_radius / 10:.6f} nm\n"
+                f"  norm1: {self.norm1}\n"
+                f"  norm2: {self.norm2}")
+    
+    def __eq__(self, other):
+        if not isinstance(other, ReactionTemplate):
+            return False
+        return self.expression == other.expression
