+++ ionerdss/model/pdb/core.py
@@ -10,9 +10,9 @@ import os
 from . import io
 from .coarse_grain import coarse_grain_structure
 from .homolog_detection import identify_homologous_chains
-from .homolog_alignment import regularize_structure
-from .reaction import build_reactions, rescale_reaction_energies
-from .visualize import plot_structure, save_structure_outputs
+from .homolog_alignment import regularize_model
+from .reaction import build_binding_reactions
+from .visualize import 
 from ..components import Model  # assuming you have a base Model class
 
 class PDBModel(Model):
